+++ src/grepros/common.py
@@ -9,13 +9,14 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     23.10.2021
-@modified    31.10.2021
+@modified    01.11.2021
 ------------------------------------------------------------------------------
 """
 from __future__ import print_function
 try: import builtins  # Py3
 except ImportError: import __builtin__ as builtins  # Py2
 import datetime
+import hashlib
 import glob
 import math
 import os
@@ -414,6 +415,36 @@ def make_live_time(stamp):
         stamp, sign = float(stamp), ("+" == stamp[0] if stamp[0] in "+-" else None)
         shift = 0 if sign is None else time.time()
     return rospy.Time(stamp + shift)
+
+
+def make_message_hash(msg, include=(), exclude=()):
+    """
+    Returns hashcode for ROS message, as a hex digest.
+
+    @param   include   message fields to include if not all, as [((nested, path), re.Pattern())]
+    @param   exclude   message fields to exclude, as [((nested, path), re.Pattern())]
+    """
+    scalar = lambda n: n[:n.index("[")] if "[" in n else n  # Returns type from type[..]
+    hasher = hashlib.md5()
+
+    def walk_message(obj, top=()):
+        fieldmap = get_message_fields(obj)
+        fieldmap = filter_fields(fieldmap, (), include=include, exclude=exclude)
+        for k, t in fieldmap.items():
+            v, path = get_message_value(obj, k, t), top + (k, )
+            if hasattr(v, "__slots__"):
+                walk_message(v, path)
+            elif isinstance(v, (list, tuple)) and scalar(t) not in ROS_BUILTIN_TYPES:
+                for x in v: walk_message(x, path)
+            else:
+                s = "%s=%s" % (path, v)
+                hasher.update(s.encode("utf-8", errors="backslashreplace"))
+        if not hasattr(obj, "__slots__"):
+            s = "%s=%s" % (top, obj)
+            hasher.update(s.encode("utf-8", errors="backslashreplace"))
+
+    walk_message(msg)
+    return hasher.hexdigest()
 
 
 def merge_spans(spans):