+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ Changelog for package grepros
 
 0.1.1 (2021-11-01)
 -------------------
-* rename arguments --noselect-field and --noprint-field to --no-select-field and --no-print-field
+* add --unique-only option
+* rename options --noselect-field and --noprint-field to --no-select-field and --no-print-field
 
 0.1.0 (2021-10-31)
 -------------------